face-health-backend/requirements.txt
flask
gunicorn
flask-cors
google-generativeai